requirements.txt
audiogen==0.1.2
certifi==2019.9.11
chardet==3.0.4
falcon==2.0.0
gunicorn==19.9.0
idna==2.8
lxml==4.4.1
numpy==1.17.2
pyparsing==2.4.2
python-dotenv==0.10.3
redis==3.3.8
requests==2.22.0
scipy==1.3.1
git+https://github.com/wireman27/svgpathtools.git#egg=svgpathtools
svgwrite==1.3.1
tones==1.0.0
urllib3==1.25.6
